requirements.txt
numpy==1.22.3
pillow==9.4.0
opencv-python==4.6.0.66
scikit-image==0.19.3
scikit-learn==1.1.1
monai>=0.9
nibabel
Pillow
gdown
torchVision
tqdm
psutil
pandas
einops